requirements/extra_apps.txt
django-taggit
django-staticfiles
django-haystack
django-authority